* Fixed all compiler warnings (DIRMINA-274) Remove Java Generics Compiler Warnings * Applied Java 5 Generics to ** Collections ** Sets ** Maps ** Lists ** Queues ** Hashtables
git-svn-id: https://svn.apache.org/repos/asf/mina/branches/1.1@525038 13f79535-47bb-0310-9956-ffa450edef68

+++ integration-jmx/src/main/java/org/apache/mina/integration/jmx/IoServiceManager.java
@@ -25,7 +25,6 @@ import java.util.Iterator;
 
 import org.apache.mina.common.IoService;
 import org.apache.mina.common.IoSession;
-import org.apache.mina.management.IoSessionStat;
 import org.apache.mina.management.StatCollector;
 
 
@@ -107,25 +106,25 @@ public class IoServiceManager implements IoServiceManagerMBean
 
     public float getAverageByteReadThroughput()
     {
-        return collector.getBytesReadThroughput() / (float) collector.getSessionCount();
+        return collector.getBytesReadThroughput() / collector.getSessionCount();
     }
 
 
     public float getAverageByteWrittenThroughput()
     {
-        return collector.getBytesWrittenThroughput() / (float) collector.getSessionCount();    
+        return collector.getBytesWrittenThroughput() / collector.getSessionCount();    
     }
 
 
     public float getAverageMessageReadThroughput()
     {
-        return collector.getMsgReadThroughput() / (float) collector.getSessionCount();
+        return collector.getMsgReadThroughput() / collector.getSessionCount();
     }
 
 
     public float getAverageMessageWrittenThroughput()
     {
-        return collector.getMsgWrittenThroughput() / (float) collector.getSessionCount();    
+        return collector.getMsgWrittenThroughput() / collector.getSessionCount();    
     }
